+++ lib/facil/cli/fio_cli.c
@@ -299,9 +299,18 @@ print_help:
   exit(0);
 }
 
+static void fio_cli_end_promise(void *ignr_) {
+  /* make sure fio_cli_end is called before facil.io exists. */
+  fio_cli_end();
+  (void)ignr_;
+}
+
 void fio_cli_start AVOID_MACRO(int argc, char const *argv[], int unnamed_min,
                                int unnamed_max, char const *description,
                                char const **names) {
+  static fio_lock_i run_once = FIO_LOCK_INIT;
+  if (!fio_trylock(&run_once))
+    fio_state_callback_add(FIO_CALL_AT_EXIT, fio_cli_end_promise, NULL);
   if (unnamed_max >= 0 && unnamed_max < unnamed_min)
     unnamed_max = unnamed_min;
   fio_cli_parser_data_s parser = {